tests/wpt/tests/css/css-animations/parsing/animation-fill-mode-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-fill-mode with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-fill-mode">
<meta name="assert" content="animation-fill-mode supports the full grammar '<single-animation-fill-mode> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("animation-fill-mode", "none");
test_valid_value("animation-fill-mode", "forwards");
test_valid_value("animation-fill-mode", "backwards");
test_valid_value("animation-fill-mode", "both");
test_valid_value("animation-fill-mode", "none, forwards, backwards, both");
</script>
</body>
</html>
